backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
tensorflow 
python-jose[cryptography]
passlib[bcrypt] 